renaming cloudwatch module to cloudwatch_destinations
