+++ streamalert_cli/terraform/cloudwatch_destinations.py
