+++ streamalert_cli/terraform/generate.py
@@ -29,7 +29,9 @@ from streamalert_cli.terraform.alert_processor import generate_alert_processor
 from streamalert_cli.terraform.apps import generate_apps
 from streamalert_cli.terraform.athena import generate_athena
 from streamalert_cli.terraform.cloudtrail import generate_cloudtrail
-from streamalert_cli.terraform.cloudwatch import generate_cloudwatch_destinations
+from streamalert_cli.terraform.cloudwatch_destinations import (
+    generate_cloudwatch_destinations,
+)
 from streamalert_cli.terraform.firehose import generate_firehose
 from streamalert_cli.terraform.flow_logs import generate_flow_logs
 from streamalert_cli.terraform.helpers import terraform_check
